Refactored kf_wrapper and relocated to kubernetes folder
