+++ kubernetes/kfserving_wrapper/TSModelRepository.py
@@ -1,0 +1,30 @@
+import os
+from kfserving.kfmodel_repository import KFModelRepository, MODEL_MOUNT_DIRS
+from TorchserveModel import TorchserveModel
+import logging
+
+logging.basicConfig(level=kfserving.constants.KFSERVING_LOGLEVEL)
+
+
+class TSModelRepository(KFModelRepository):
+    """The class takes in the KFServing side model related attributes.
+    Args:
+        KFModelRepository (object): The parameters from the KFModelRepository is passed
+        as inputs to the TSModel Repository.
+    """
+
+    def __init__(self, inference_address: str, management_address: str, model_dir: str):
+        """The Inference Address, Management Address and the Model Directory from the KFServing
+        side is initialized here.
+
+        Args:
+            inference_address (str): The Inference Address present in the KFServing side.
+            management_address (str): The Management Address present in the kfserving side.
+            model_dir (str): the directory of the model artefacts in the kfserving side.
+        """
+        super().__init__(model_dir)
+        logging.info("TSModelRepo is initialized")
+        self.inference_address = inference_address
+        self.management_address = management_address
+        self.model_dir = model_dir
+
