--- ts/kfserving_wrapper/config_parse/config.properties
@@ -1,8 +1,0 @@
-inference_address=http://0.0.0.0:8085
-management_address=http://0.0.0.0:8081
-NUM_WORKERS=1
-number_of_netty_threads=4
-job_queue_size=10
-service_envelope=kfserving
-model_store=/home/model-server/shared/model-store
-model_snapshot={"name":"startup.cfg","modelCount":1,"models":{"mnist":{"1.0":{"defaultVersion":true,"marName":"mnist.mar","minWorkers":1,"maxWorkers":5,"batchSize":5,"maxBatchDelay":200,"responseTimeout":60}}}}
