+++ ts/torch_handler/vision_handler.py
@@ -10,6 +10,7 @@ from PIL import Image
 from .base_handler import BaseHandler
 from captum.attr import IntegratedGradients
 
+
 class VisionHandler(BaseHandler, ABC):
     """
     Base class for all vision handlers
@@ -39,3 +40,4 @@ class VisionHandler(BaseHandler, ABC):
     def get_insights(self, tensor_data, raw_data, target=0):
         print("input shape", tensor_data.shape)
         return self.ig.attribute(tensor_data, target=target, n_steps=15).tolist()
+
